+++ packages/language/src/cli.ts
@@ -1,6 +1,6 @@
 import { readFileSync } from "fs";
 import { exit } from "process";
-import { generate } from "./generate.js";
+import { generate } from "./ecchi-generator.js";
 
 const input = readFileSync(process.argv[process.argv.length-1], 'utf8');
 generate(input).then(dts => console.log(dts)).catch(e => {
